+++ out/lib/reporters/console.js
@@ -15,12 +15,6 @@
 
     ConsoleReporter.prototype.config = null;
 
-    ConsoleReporter.prototype.passed = 0;
-
-    ConsoleReporter.prototype.failed = 0;
-
-    ConsoleReporter.prototype.total = 0;
-
     function ConsoleReporter(config) {
       var _base, _base1, _base2, _base3, _base4, _base5, _base6, _ref, _ref1, _ref2, _ref3, _ref4, _ref5, _ref6;
       this.errors || (this.errors = []);
@@ -44,7 +38,7 @@
         _base5.summaryPass = "%s/%s tests ran successfully, everything passed";
       }
       if ((_ref6 = (_base6 = this.config).summaryFail) == null) {
-        _base6.summaryFail = "%s/%s tests ran successfully, with %s errors";
+        _base6.summaryFail = "FAILURE: %s/%s tests ran successfully; %s failed, %s incomplete, %s errors";
       }
       if (cliColor != null) {
         if (__indexOf.call(typeof process !== "undefined" && process !== null ? process.argv : void 0, '--no-colors') < 0) {
@@ -59,28 +53,30 @@
     }
 
     ConsoleReporter.prototype.getSuiteName = function(suite) {
-      var parentSuite, parentSuiteName, suiteName;
-      suiteName = suite.getSuiteName();
-      parentSuite = suite.getParentSuite();
-      if (parentSuite) {
-        parentSuiteName = this.getSuiteName(parentSuite);
-        suiteName = "" + parentSuiteName + this.config.sub + suiteName;
-      }
-      return suiteName;
+      return this.joe.getSuiteName(suite, this.config.sub);
     };
 
     ConsoleReporter.prototype.getTestName = function(suite, testName) {
-      var suiteName;
+      var result, suiteName;
+      result = '';
       if (suite != null) {
         suiteName = this.getSuiteName(suite);
-        testName = "" + suiteName + this.config.sub + testName;
+        result += "" + suiteName;
+        if (testName) {
+          result += "" + this.config.sub + testName;
+        }
+      } else {
+        result = testName;
       }
-      return testName;
+      return result;
     };
 
     ConsoleReporter.prototype.startSuite = function(suite) {
       var message, suiteName;
       suiteName = this.getSuiteName(suite);
+      if (!suiteName) {
+        return this;
+      }
       message = "" + suiteName + this.config.start;
       console.log(message);
       return this;
@@ -88,10 +84,10 @@
 
     ConsoleReporter.prototype.finishSuite = function(suite, err) {
       var check, message, suiteName;
-      if (err && this.errors.length === 0) {
-        this.uncaughtException(err);
-      }
       suiteName = this.getSuiteName(suite);
+      if (!suiteName) {
+        return this;
+      }
       check = (err ? this.config.fail : this.config.pass);
       message = "" + suiteName + check;
       console.log(message);
@@ -100,8 +96,10 @@
 
     ConsoleReporter.prototype.startTest = function(suite, testName) {
       var message;
-      ++this.total;
       testName = this.getTestName(suite, testName);
+      if (!testName) {
+        return this;
+      }
       message = "" + testName + this.config.start;
       console.log(message);
       return this;
@@ -109,47 +107,33 @@
 
     ConsoleReporter.prototype.finishTest = function(suite, testName, err) {
       var check, message;
-      if (err) {
-        this.errors.push({
-          suite: suite,
-          testName: testName,
-          err: err
-        });
-        ++this.failed;
-      } else {
-        ++this.passed;
-      }
       testName = this.getTestName(suite, testName);
+      if (!testName) {
+        return this;
+      }
       check = (err ? this.config.fail : this.config.pass);
       message = "" + testName + check;
       console.log(message, (typeof process !== "undefined" && process !== null) === false && err ? [err, err.stack] : '');
       return this;
     };
 
-    ConsoleReporter.prototype.uncaughtException = function(err) {
-      this.errors.push({
-        testName: 'uncaughtException',
-        err: err
-      });
-      return this;
-    };
-
-    ConsoleReporter.prototype.exit = function() {
-      var error, index, suite, testName, _i, _len, _ref, _ref1;
-      if (this.errors.length === 0) {
-        console.log("\n" + this.config.summaryPass, this.passed, this.total);
-      } else {
-        console.log("\n" + this.config.summaryFail, this.passed, this.total, this.errors.length);
-        _ref = this.errors;
-        for (index = _i = 0, _len = _ref.length; _i < _len; index = ++_i) {
-          error = _ref[index];
-          suite = error.suite, testName = error.testName, err = error.err;
+    ConsoleReporter.prototype.exit = function(exitCode) {
+      var errorLog, errorLogs, index, suite, testName, totalErrors, totalFailedTests, totalIncompleteTests, totalPassedTests, totalTests, _i, _len, _ref, _ref1;
+      _ref = this.joe.getTotals(), totalTests = _ref.totalTests, totalPassedTests = _ref.totalPassedTests, totalFailedTests = _ref.totalFailedTests, totalIncompleteTests = _ref.totalIncompleteTests, totalErrors = _ref.totalErrors;
+      if (exitCode) {
+        errorLogs = this.joe.getErrorLogs();
+        console.log("\n" + this.config.summaryFail, totalPassedTests, totalTests, totalFailedTests, totalIncompleteTests, totalErrors);
+        for (index = _i = 0, _len = errorLogs.length; _i < _len; index = ++_i) {
+          errorLog = errorLogs[index];
+          suite = errorLog.suite, testName = errorLog.testName, err = errorLog.err;
           testName = this.getTestName(suite, testName);
           console.log("\n" + this.config.failHeading, index + 1);
           console.log(testName);
           console.log(((_ref1 = err.stack) != null ? _ref1.toString() : void 0) || err);
         }
         console.log('');
+      } else {
+        console.log("\n" + this.config.summaryPass, totalPassedTests, totalTests);
       }
       return this;
     };