+++ out/test/joe.test.js
@@ -8,7 +8,7 @@
 
   everythingTestPath = __dirname + '/../example/everything.test.js';
 
-  expected = "10/11 tests ran successfully, with 1 errors";
+  expected = "FAILURE: 10/11 tests ran successfully; 1 failed, 0 incomplete, 1 errors";
 
   stdout = '';
 
@@ -24,11 +24,14 @@
   });
 
   runner.on('exit', function(code) {
-    var message, result;
-    result = stdout.indexOf(expected) !== -1;
-    message = 'THE ABOVE WAS EXACTLY WHAT WE EXPECTED. TESTS HAVE PASSED!';
-    assert.ok(result, message);
-    return console.log(message);
+    var pass;
+    pass = stdout.indexOf(expected) !== -1;
+    if (pass) {
+      console.log('THE ABOVE IS WHAT WE EXPECTED. TESTS HAVE PASSED');
+    } else {
+      console.log('THE ABOVE IS NOT WHAT WE WE EXPECTED. TESTS HAVE FAILED');
+    }
+    return assert.ok(pass);
   });
 
 }).call(this);